[Address] Pr 수정 - MapActivity, AddressActivity ViewMdoel 통합 - AddressActivity layout 수정
