+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,6 @@
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
     <uses-permission android:name="android.permission.CHANGE_WIFI_STATE" />
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_BACKGROUND_LOCATION" />
 
 
@@ -18,7 +17,7 @@
             android:supportsRtl="true"
             android:theme="@style/Theme.Sticky">
         <activity android:name=".ui.address.SetAddressActivity" />
-        <activity android:name=".ui.map.MapActivity" />
+        <activity android:name=".ui.address.MapActivity" />
         <activity
                 android:name=".ui.share.ShareActivity"
                 android:screenOrientation="portrait" />
