+++ app/src/main/java/com/nexters/sticky/ui/SplashActivity.kt
@@ -9,7 +9,6 @@ import androidx.lifecycle.lifecycleScope
 import com.nexters.sticky.R
 import com.nexters.sticky.ui.address.SetAddressActivity
 import com.nexters.sticky.ui.gps.NeedPermissionActivity
-import com.nexters.sticky.ui.main.MainActivity
 import com.nexters.sticky.utils.LOCATION_PERMISSION_REQUEST_CODE
 import com.nexters.sticky.utils.checkLocationPermission
 import com.nexters.sticky.utils.requestLocationPermission
