--- app/src/main/java/com/nexters/sticky/ui/map/MapViewModel.kt
@@ -1,13 +1,0 @@
-package com.nexters.sticky.ui.map
-
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class MapViewModel : ViewModel() {
-	val text = MutableLiveData<String>()
-//    var sampleList: LiveData<MutableList<Sample>> = MutableLiveData()
-
-	fun setText(value: String) {
-		text.value = value
-	}
-}
