+++ app/src/main/res/drawable/icon_search.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="130"
+    android:viewportWidth="130" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000000" android:fillType="evenOdd" android:pathData="M86.333,54.333C86.333,72.006 72.007,86.333 54.333,86.333C36.66,86.333 22.333,72.006 22.333,54.333C22.333,36.66 36.66,22.333 54.333,22.333C72.007,22.333 86.333,36.66 86.333,54.333ZM80.497,88.039C73.272,93.655 64.193,97 54.333,97C30.769,97 11.667,77.897 11.667,54.333C11.667,30.769 30.769,11.667 54.333,11.667C77.898,11.667 97,30.769 97,54.333C97,64.193 93.656,73.272 88.039,80.497L111.438,103.895C113.521,105.978 113.521,109.355 111.438,111.438C109.355,113.521 105.978,113.521 103.896,111.438L80.497,88.039Z"/>
+</vector>
